Rename S3 to AWS_S3
