+++ homeassistant/components/aws_s3/quality_scale.yaml
