+++ homeassistant/components/aws_s3/strings.json
@@ -9,18 +9,18 @@
           "endpoint_url": "Endpoint URL"
         },
         "data_description": {
-          "access_key_id": "Access key ID to connect to S3 API",
-          "secret_access_key": "Secret access key to connect to S3 API",
+          "access_key_id": "Access key ID to connect to AWS S3 API",
+          "secret_access_key": "Secret access key to connect to AWS S3 API",
           "bucket": "Bucket must already exist and be writable by the provided credentials.",
           "endpoint_url": "Endpoint URL provided to [Boto3 Session]({boto3_docs_url}). Region-specific [AWS S3 endpoints]({aws_s3_docs_url}) are available in their docs."
         },
-        "title": "Add S3 bucket"
+        "title": "Add AWS S3 bucket"
       }
     },
     "error": {
-      "cannot_connect": "[%key:component::s3::exceptions::cannot_connect::message%]",
-      "invalid_bucket_name": "[%key:component::s3::exceptions::invalid_bucket_name::message%]",
-      "invalid_credentials": "[%key:component::s3::exceptions::invalid_credentials::message%]",
+      "cannot_connect": "[%key:component::aws_s3::exceptions::cannot_connect::message%]",
+      "invalid_bucket_name": "[%key:component::aws_s3::exceptions::invalid_bucket_name::message%]",
+      "invalid_credentials": "[%key:component::aws_s3::exceptions::invalid_credentials::message%]",
       "invalid_endpoint_url": "Invalid endpoint URL"
     },
     "abort": {
